Destroy Reservation and Home Page
Create tests (unit and integration), controller action, and model
callback to handle destroying a reservation. In particular, whenever
a reservation is destroyed, all lockers that belong to that
reservation will be unassigned from the reservation (allowing them
to be assigned to another reservation in the future.)

Create tests, routes, controller, and view for the Home page.